user-service/requirements.txt

grpcio==1.60.0
grpcio-tools==1.60.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.9
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.8.0